refactor(project): go-style v2 to pythonic v3
