+++ README.md
@@ -2,6 +2,14 @@
 
 This is an example of implementing a Pokémon API based on the Clean Architecture in a Python project, referencing [**go-clean-arch**](https://github.com/bxcodec/go-clean-arch)
 
+## Changelog
+
+- **v1**: checkout to the [v1 branch ](https://github.com/cdddg/py-clean-arch/tree/v1)<br>
+  archived to v1 branch on 2021/04 <br>desc: Initial proposal by me
+- **v2**: checkout to the [v2 branch ](https://github.com/cdddg/py-clean-arch/tree/v2)<br>archived to v1 branch on 2023/07 <br>
+  desc: Improvement from v1, the details can be seen on this PR [#1](https://github.com/cdddg/py-clean-arch/pull/1)
+- **v3**: [master branch](https://github.com/cdddg/py-clean-arch/tree/master) <br>merged to master on 2023/08. <br>desc: Transitioned from Go-centric design to conventional Python practices. Details available in this PR [#11](https://github.com/cdddg/py-clean-arch/pull/11).https://github.com/cdddg/py-clean-arch/pull/11)
+
 ## Description
 
 Rule of Clean Architecture by Uncle Bob [^1]
@@ -22,9 +30,36 @@ The project, like the original project, has 4 domain layers:
 In addition, Add [Unit of Work Pattern](./src/settings/unit_of_work.py) [^2], [Dependency Injection Pattern](./src/settings/dependency_injection.py) [^3] <br>
 and use Asynchronous SQLalchemy[^4]
 
-#### The diagram:
+### Project Structure Overview & Clean Architecture Mapping
 
-<img src="./docs/clean-arch.png">
+The directory structure below provides a high-level view of the project. Each directory or module corresponds to a layer in Uncle Bob's Clean Architecture. Please note that only the main directories and key files are listed here. Some specific files or subdirectories might be omitted to highlight the overall architecture and the primary purpose of each layer.
+
+```
+src
+├── app/                                - Directory for business logic.
+│   ├── deliveries/                     - Interface layer for external interactions. 
+│   │                                   ("Frameworks and Drivers" Layer in Clean Architecture)
+│   │
+│   ├── usecases/                       - Application services layer representing business use cases.
+│   │                                   ("Use Cases" Layer in Clean Architecture)
+│   │
+│   ├── repositories/                   - Data storage and interaction module. 
+│   │                                   ("Interface" Adapters Layer in Clean Architecture)
+│   │
+│   └── di/                             - Dependency injection module.
+│
+├── models/                             - Domain models representing core business logic.
+│                                       ("Entities" Layer in Clean Architecture)
+│
+├── common/                             - Module containing shared code and utilities. 
+├── settings/                           - Application configurations. 
+├── tests/                              - Testing module for the application. 
+└── main.py                             - Main file to launch the application.
+```
+
+![clean-arch-01](./docs/clean-arch-01.png)
+
+<img src="./docs/clean-arch-02.png" alt="clean-arch-02" style="zoom:75%;" />
 
 ### How To Run This Project
 
@@ -101,17 +136,8 @@ To generate a coverage report:
 $ pytest --cov
 ```
 
-## Changelog
-
-- **v1**: checkout to the [v1 branch ](https://github.com/cdddg/py-clean-arch/tree/v1)<br>
-  archived to v1 branch on 2018 <br>
-  Desc: Initial proposal by me
-- **v2**: master branch <br>
-  merged to master on 2023. <br>
-  Desc: Improvement from v1, the details can be seen on this PR [#1](https://github.com/cdddg/py-clean-arch/issues/1)
-
 [^1]: https://github.com/bxcodec/go-clean-arch#description
 [^2]: https://www.cosmicpython.com/book/chapter_06_uow.html
 [^3]: https://en.wikipedia.org/wiki/Dependency_injection
 [^4]:  The asyncio extension as of SQLAlchemy 1.4.3 can now be considered to be **beta level** software. API details are subject to change however at this point it is unlikely for there to be significant backwards-incompatible changes. https://docs.sqlalchemy.org/en/14/orm/extensions/asyncio.html
-[^5]: Test results as of [today's date, e.g., August 12, 2023].
+
