+++ src/app/deliveries/graphql/pokomon/query.py
@@ -1,7 +1,7 @@
 import strawberry
 from strawberry.types import Info
 
-import pkg.usecases.pokemon as pokemon_ucase
+import app.usecases.pokemon as pokemon_ucase
 
 from .schema import PokemonNode
