+++ src/app/deliveries/graphql/pokomon/scalar.py
@@ -1,6 +1,6 @@
 import strawberry
 
-from core.type import UUIDStr
+from common.type import UUIDStr
 
 
 @strawberry.scalar(description=UUIDStr.__doc__)
